use consistent memory formatting
